Updated interfaces
